Reduce visibility of more runtime methods
This is a large refactoring as classes around PyType are (necessarily)
designed together.
